# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PaperGo headless DMS",
    "version": "0.4.0",
    "description": "One organization per SQLite deployment. Immutable schema/content revisions, SharePoint-style exclusive permission scopes, head/published projections, rich typed fields, reusable templates, saved queries, taxonomy, and typed relationships. Libraries with webdav_enabled are also served over WebDAV (RFC 4918 class 1 and 2) at /webdav/{libraryID}/, outside this contract; WebDAV accepts these bearer tokens or HTTP Basic with a WebDAV credential password."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/v1/workspaces": {
      "get": {
        "summary": "List visible workspaces",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Resource"
                      }
                    },
                    "next_cursor": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "parameters": [
          {
            "name": "after",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          }
        ]
      },
      "post": {
        "summary": "Create a workspace with creator manage permission",
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Resource"
                }
              }
            },
            "headers": {
              "ETag": {
                "description": "Quoted numeric resource lock version. Field updates use the container version.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255
                  },
                  "tags": {
                    "type": "array",
                    "maxItems": 50,
                    "uniqueItems": true,
                    "items": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 64
                    }
                  }
                },
                "required": [
                  "name"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/v1/resources": {
      "get": {
        "summary": "Browse/search a workspace or parent scope",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Resource"
                      }
                    },
                    "next_cursor": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "parameters": [
          {
            "name": "after",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          },
          {
            "name": "workspace_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "parent_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "schema": {
              "default": "auto",
              "enum": [
                "auto",
                "head",
                "published"
              ],
              "type": "string"
            },
            "name": "surface",
            "description": "Auto chooses head for draft readers and published for ordinary readers. Head requires draft access.",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^[a-z][a-z0-9_]{0,63}$"
            },
            "name": "filter_field",
            "description": "Opt-in indexed field key; requires list/library/folder parent scope.",
            "in": "query"
          },
          {
            "schema": {
              "default": "eq",
              "enum": [
                "eq",
                "gt",
                "gte",
                "lt",
                "lte"
              ],
              "type": "string"
            },
            "name": "filter_op",
            "description": "Typed comparison. Boolean supports eq only.",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "name": "filter_value",
            "description": "Typed value, parsed without floating-point conversion for integer and decimal fields.",
            "in": "query"
          }
        ]
      }
    },
    "/v1/resources/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "summary": "Read resource metadata",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Resource"
                }
              }
            },
            "headers": {
              "ETag": {
                "description": "Quoted numeric resource lock version. Field updates use the container version.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "parameters": [
          {
            "schema": {
              "default": "auto",
              "enum": [
                "auto",
                "head",
                "published"
              ],
              "type": "string"
            },
            "name": "surface",
            "description": "Auto chooses head for draft readers and published for ordinary readers. Head requires draft access.",
            "in": "query"
          }
        ]
      },
      "patch": {
        "summary": "Update resource; values replaces the whole custom-value object",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Resource"
                }
              }
            },
            "headers": {
              "ETag": {
                "description": "Quoted numeric resource lock version. Field updates use the container version.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "parameters": [
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "description": "Current quoted numeric resource ETag, e.g. \"1\".",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateResource"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a folder or item with everything below it (write required on all of it)",
        "description": "Deleted resources become retained tombstones: they disappear from every read, search, query and relationship, their names become free, and their revisions, publications and audit events are kept. Deletion cannot be undone.",
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "parameters": [
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "description": "Current quoted numeric resource ETag, e.g. \"1\".",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/v1/webdav-credentials": {
      "get": {
        "summary": "List the caller's WebDAV credentials",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "maxItems": 20,
                      "items": {
                        "$ref": "#/components/schemas/WebDAVCredential"
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      },
      "post": {
        "summary": "Issue a WebDAV app password for the caller (at most 20)",
        "description": "Clients such as Windows Explorer authenticate to /webdav/ with HTTP Basic: any user name and this password. The password is returned only in this response; only its SHA-256 is stored.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebDAVCredentialInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewWebDAVCredential"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/v1/webdav-credentials/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "delete": {
        "summary": "Revoke one of the caller's WebDAV credentials",
        "responses": {
          "204": {
            "description": "Revoked"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
    "/v1/resources/{id}/fields": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "summary": "Read list/library field definitions",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/FieldDefinition"
                      }
                    }
                  },
                  "required": [],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      },
      "post": {
        "summary": "Add a field definition (manage required)",
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FieldDefinition"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateField"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "Problem": {
        "description": "RFC 9457 problem response. Possible status codes: 400, 401, 403, 404, 409, 413, 415, 422, 428, 500, 503.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      }
    },
    "schemas": {
...
      "UpdateResource": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "tags": {
            "type": "array",
            "maxItems": 50,
            "uniqueItems": true,
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64
            }
          },
          "values": {
            "type": "object",
            "additionalProperties": true
          },
          "publishing_enabled": {
            "default": false,
            "description": "Lists/libraries only. False publishes content changes immediately. Changing policy requires manage access; disabling publishes all current heads atomically.",
            "type": "boolean"
          },
          "webdav_enabled": {
            "description": "Libraries only; changing it requires manage access.",
            "type": "boolean"
          },
          "parent_id": {
            "description": "Moves a folder or item below another folder of the same list or library, or to its top level. Requires write access to the new parent. A move alone creates no content revision; inheriting resources take the new parent's permission scope.",
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [],
        "additionalProperties": false
      },
...
    }
  }
}